32/64 bit changes to Epetra, FEApp, stokhos, phalanx.
Include Epetra_ConfigDefs.h before any other file in those
Epetra files that make use of EPETRA_NO_64BIT_GLOBAL_INDICES
or EPETRA_NO_32BIT_GLOBAL_INDICES macros.  Avoids vtable
inconsistency (at least) across translation units.

Remove the recently introduced unsigned int/long long based
constructors and functions from Map classes.  They were
introduced to allow overloading in external packages so that
ambiguity between int and long long does not happen but they
lead to a different overloading ambiguity when code is compiled
in single mode (32 or 64).  The correct fix is below.

Cast argument to Epetra_LocalMap to be always int.  It will
then create int map when 32 bit indices are present and long
long map when they are not.  This avoids ambiguity between
int and long long when size_t or unsigned int is passed.

Cast argument to LID() to be int or long long depending on
compilation mode.

Keep some long long code that was always available inside
but trying to reduce clutter in object code and so that
linker/compiler errors can happen when they should.

Explicitly exclude the Epetra_LongLong*/ h/cpp files when
long long indices are not present even though the
translation unit would be empty.  Reduce cmake clutter.

When a class contains foo_int_ and foo_LL_ data members,
make them available depending on compilation mode.  Saves
sizeof(object) and hopefully leads to linker/compiler errors
when they should happen.

+++ demos/FEApp/example/nox_stokhos/hns_jacobian_KL.cpp
@@ -284,7 +284,7 @@ int main(int argc, char *argv[]) {
     Teuchos::RCP<const Epetra_Comm> stoch_comm =
       sg_parallel_data->getStochasticComm();
     Teuchos::RCP<const Epetra_BlockMap> stoch_overlap_map = 
-      Teuchos::rcp(new Epetra_LocalMap(sz, 0, *stoch_comm));
+      Teuchos::rcp(new Epetra_LocalMap(static_cast<int>(sz), 0, *stoch_comm));
 
     // Stochastic parameters
      parameterParams.set("Number of Parameter Vectors", 2);
@@ -473,7 +473,7 @@ int main(int argc, char *argv[]) {
     Teuchos::RCP<const Epetra_BlockMap> kl_stoch_row_map =
       kl_epetraCijk->getStochasticRowMap();
     Teuchos::RCP<const Epetra_BlockMap> kl_ov_stoch_row_map = 
-      Teuchos::rcp(new Epetra_LocalMap(sz2, 0, kl_comm->TimeDomainComm()));
+      Teuchos::rcp(new Epetra_LocalMap(static_cast<int>(sz2), 0, kl_comm->TimeDomainComm()));
     
     Teuchos::RCP<const Epetra_Map> base_x_map = model->get_x_map();
     Teuchos::RCP<const Epetra_Map> base_f_map = model->get_f_map();
